requirements.txt
pytest==8.3.5
allure-pytest
selene==2.0.0rc9
python-dotenv==0.20.0